pyproject.toml
[project]
name = "deepseek-textbook"
version = "0.1.0"
description = "Generate textbook content using DeepSeek API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pypandoc>=1.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
